% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_titre_image.R
\name{ajouter_titre_image}
\alias{ajouter_titre_image}
\title{On ajoute un titre sur une image}
\usage{
ajouter_titre_image(titre, nom_image)
}
\arguments{
\item{titre}{Titre qu'on souhaite ajouter}

\item{nom_image}{Nom de l'image (qui doit être préchargée au bon format, on ne charge pas un png ou jpg)}
}
\value{
une image avec un titre
}
\description{
Concrètement, c'est surtout utilisé pour le moment pour ajouter le nom des DOM en haut de l'image
D'ailleurs, ni le positionnement, ni la taille, ni quoique ce soit ne sont proposé en paramètre, preuve du côté très limité pour le moment de cette fonction
}
