% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appliquer_filtrage_liste.R
\name{appliquer_filtrage_liste}
\alias{appliquer_filtrage_liste}
\title{Appliquer filtrage à une liste}
\usage{
appliquer_filtrage_liste(
  liste_tableau,
  seuil_ecart = 0.5,
  seuil_niveau = 100,
  seuil_na = 0.5,
  affichage = FALSE
)
}
\arguments{
\item{liste_tableau}{liste de tableaux}

\item{seuil_ecart}{Le seuil de filtrage, en \%. Par exemple, un seuil à 0.5 signifie qu'on va garder les valeurs qui sont à moins de 50\% d'écart de la moyenne (c'est le seuil par défaut si rien n'est précisé)}

\item{seuil_niveau}{le niveau à partir duquel on commence une correction. Par défaut égal à 100 : cela signifie que si la moyenne est inférieure à 100, on ne fait rien}

\item{seuil_na}{s'il y a trop de valeurs en NA dans la ligne, on préfère ne pas calculer la moyenne, et donc ne pas faire de correction (car la moyenne ne veut rien dire). Ce paramètre permet de fixer un seuil (par défaut, 50\% : s'il manque plus de 50\% de valeurs sur la ligne, on ne change rien). Ce seuil doit être défini en fonction du nombre de valeurs probablement}

\item{affichage}{booléen qui indique si on veut afficher un diagnostic ou non}
}
\value{
la même liste de tableaux, mais filtrée
}
\description{
Cette fonction applique le filtrage des données (remplacer par NA) à une liste de tableaux
}
