% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_mode.R
\name{calculer_mode}
\alias{calculer_mode}
\title{Fonction calculer mode}
\usage{
calculer_mode(x)
}
\arguments{
\item{x}{un vecteur de données}
}
\value{
Un élément avec
#' @examples
calculer_mode(c(1, 2, 2, 3, 4))
}
\description{
Cette fonction calcule la modalité la plus fréquente pour un vecteur de données. Si on lui fournit un argument vide, il ne renvoie rien
}
