% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_france.R
\name{carte_france}
\alias{carte_france}
\title{Cette fonction permet, à partir d'un data frame et/ou d'un fichier Excel, de créer une carte de France par département}
\usage{
carte_france(dataframe, variable, nom_legende, breaks, labels)
}
\arguments{
\item{dataframe}{Une tableau de données, ou un fichier Excel. Ce dernier doit comporter des données départementales (une ligne par département), et au moins deux colonnes (une pour les départements, une pour la variable souhaitée)}

\item{variable}{La variable sur laquelle on souhaite faire une carte}

\item{nom_legende}{La légende souhaitée}

\item{breaks}{Un vecteur avec les intervalles proposés}

\item{labels}{Les labels des intervalles}
}
\value{
Une carte de France telle que souhaitée
}
\description{
Cette carte de France est réalisée à partir d'une variable continue, pour laquelle il va falloir définir des "classes",
C'est à dire des intervalles
}
