% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_restreindre_base.R
\name{carte_restreindre_base}
\alias{carte_restreindre_base}
\title{Restriction de la base des cartes en fonction des régions sélectionnées}
\usage{
carte_restreindre_base(donnees, region)
}
\arguments{
\item{donnees}{des données issues du shapefile}

\item{region}{la région sélectionnée (parmi les régions de France + FRANCEMETRO + FRANCEENTIERE)}
}
\value{
le tableau, filtré comme il faut
}
\description{
Fonction assez simple, a priori fonction intermédiaire du package (mais laissée en export pour le moment)
}
