% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_statique.R
\name{carte_statique}
\alias{carte_statique}
\title{Fonction principale pour faire une carte statique à l'échelle départementale.}
\usage{
carte_statique(
  donnees,
  var,
  niveau_geo = "France_entiere",
  choro = TRUE,
  discretisation = "quantile",
  nbre_classes = 5,
  titre_legende = "",
  titre_carte = "",
  source = "",
  auteur = "",
  medaillon = TRUE
)
}
\arguments{
\item{donnees}{La base de données qui contient les valeurs à afficher sur la carte. la colonne qui renseigne le département doit impérativement s'appeler 'code_insee'. Cette colonne doit être au format caractère et en avoir trois. Il ne doit pas y avoir de géométries !!}

\item{var}{Variable qui doit être affichée sur la carte.}

\item{niveau_geo}{Il y en a trois : 'France_entiere' (France métropolitaine et départements d'Outre-mer), 'Metropole' (France métropolitaine uniquement) et 'Outre-mer' (Départements d'Outre-mer uniquement).}

\item{choro}{Si TRUE (défaut) la carte sera de type choroplète, sinon elle sera en proportions.}

\item{discretisation}{Choix de la méthode de discrétisation (le détail des méthodes disponibles est accessible dans le dictionnaire de la fonction mf_get_breaks()). Utilisé uniquement avec les cartes choroplètes.}

\item{nbre_classes}{Nombre de classes dans laquelle est décomposée une variable. Utilisé uniquement avec les cartes choroplètes.}

\item{titre_legende}{Titre de la légende.}

\item{titre_carte}{Titre de la carte.}

\item{source}{Source des données.}

\item{auteur}{Auteur de la carte.}

\item{medaillon}{Si TRUE (défaut) une carte médaillon affiche l'Île-De-France.}
}
\description{
Fonction principale pour faire une carte statique à l'échelle départementale.
}
