% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsa-package.R
\docType{package}
\name{cnsa-package}
\alias{cnsa}
\alias{cnsa-package}
\title{cnsa: Fonctions utiles pour le pôle PAPS de la CNSA}
\description{
Ce package contient différentes fonctions utiles dans le cadre du travail de PAPS à la CNSA. En particulier, différentes fonctions de manipulations de données, une fonction qui permet d'inverser les "/" en copiant les noms de répertoire (très utile !), ainsi que certaines fonctions d'imputation et de traitement de la non-réponse.
}
\author{
\strong{Maintainer}: Aurélien Fortin \email{aurelien.fortin@cnsa.fr}

Other contributors:
\itemize{
  \item cre Caisse Nationale pour la Solidarité et l'Autonomie
}

}
\keyword{internal}
