% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_num.R
\name{conversion_num}
\alias{conversion_num}
\title{Convertit une entrée texte en un nombre numérique.}
\usage{
conversion_num(x)
}
\arguments{
\item{x}{Une entrée texte à convertir.}
}
\value{
Une valeur numérique correspondant à l'entrée texte.
}
\description{
Cette fonction prend en charge la conversion des entrées texte contenant des virgules,
des points-virgules, des pourcentages et des espaces, les transformant en une forme numérique appropriée.
}
\examples{
# Exemple d'utilisation:
entree_texte <- "1,234.56 \%"
# Appel de la fonction :
valeur_numerique <- conversion_num(entree_texte)
}
