% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_carte_accessible.R
\name{creation_carte_accessible}
\alias{creation_carte_accessible}
\title{Créer une carte d’accessibilité}
\usage{
creation_carte_accessible(
  donnees,
  calcul_classes = "OUI",
  classes = "quantiles",
  type_valeur = "niveau",
  couleurs = c("#E8F6FC", "#B9E4F7", "#A0DBF4", "#7EB2C9", "#5D8B9F", "#3E6677",
    "#234655", "#D1EDFA"),
  manual_pattern = c("none", "circle", "stripe", "crosshatch", "stripe", "stripe",
    "crosshatch", "stripe"),
  angle_pattern = c(0, 0, 90, 45, 180, 45, 0, 0)
)
}
\arguments{
\item{donnees}{Dataframe contenant les données par département.}

\item{calcul_classes}{Indique si les classes sont calculées automatiquement.}

\item{classes}{Type de classes à utiliser.}

\item{type_valeur}{Type de valeur affichée ("niveau" ou "pourcentage").}

\item{couleurs}{Vecteur de couleurs.}

\item{manual_pattern}{Vecteur des motifs par classe.}

\item{angle_pattern}{Vecteur des angles des motifs.}
}
\description{
Génère une carte d’accessibilité à partir de données départementales.
}
