% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_carte_indiv.R
\name{creer_carte_indiv}
\alias{creer_carte_indiv}
\title{Creer carte individuelle}
\usage{
creer_carte_indiv(donnees, region, palette, titre_legende = "Legende")
}
\arguments{
\item{donnees}{un tableau de données, format SF, avec les coordonnées}

\item{region}{la région souhaitée, incluse dans le SF (a priori FRANCEMETRO, 971, 972,...)}

\item{palette}{une palette de couleurs, fournie a priori avant (note : peut être à changer)}

\item{titre_legende}{Si la légende s'affiche : qu'est-il marqué au dessus ?}
}
\value{
une carte format leaflet
}
\description{
Cette fonction crée une carte pour une région donnée (FRANCEMETRO, ou un département d'outre mer)
}
