% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_palette.R
\name{creer_palette}
\alias{creer_palette}
\title{Cette fonction crée une palette de couleur}
\usage{
creer_palette(variable_entree, type_palette)
}
\arguments{
\item{variable_entree}{variable au format numérique ou facteur (test à l'intérieur pour adapter). Un seul vecteur par contre}

\item{type_palette}{type de palette (nom de la palette). par défaut : viridis}
}
\value{
une palette
}
\description{
L'objectif est, pour une variable donnée, de créer une palette de couleurs associée qui sera utilisée dans les légendes.
}
