% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_reference_finess.R
\name{creer_reference_finess}
\alias{creer_reference_finess}
\title{Créer référence finess}
\usage{
creer_reference_finess(origine = "GEOD", annee_ref = 2019)
}
\value{
rien
}
\description{
Crée une base de référence finess, utilisable pour du redressement, ou pour toute autre chose et l'enregistre
}
