% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_tableau_synthese.R
\name{creer_tableau_synthese}
\alias{creer_tableau_synthese}
\title{Créer un tableau synthétisant une variable}
\usage{
creer_tableau_synthese(finess_full, tab, groupe, variable, annee)
}
\arguments{
\item{finess_full}{résultat de creer_reference_finess qui contient un tableau 0-1 indiquant par année si un FINESS est ouvert}

\item{tab}{tableau avec au moins une colonne FINESS et une variable numérique à synthétiser}

\item{groupe}{catégorie selon laquelle on veut regrouper : "categetab", "code_regroup_finess", "region" ou "departement"}

\item{annee}{année étudiée sous forme de chaîne de caractères}
}
\value{
un tableau avec trois colonnes :
groupe : la liste des catégories
variable : somme des valeurs pour les FINESS ouverts pour chaque catégorie
nombre : nombre de FINESS ouverts dans chaque catégorie
}
\description{
Crée un tableau synthétisant la somme d'une variable pour une année sur tous les FINESS ouverts selon un regroupement
}
