% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumuler_cartes.R
\name{cumuler_cartes}
\alias{cumuler_cartes}
\title{Cette fonction fusionne toutes les images en une}
\usage{
cumuler_cartes(chemin = "")
}
\arguments{
\item{chemin}{argument optionnel, répertoire dans lequel on met le résultat}
}
\value{
Une carte au format jpg dans le répertoire souhaité
}
\description{
On prend en entrée toutes les cartes, format png, france métropolitaine et DOM
Et on assemble tout pour sortir un jpeg
}
