% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlever_valeurs_extremes.R
\name{enlever_valeurs_extremes}
\alias{enlever_valeurs_extremes}
\title{Enlève les valeurs extrêmes de chaque colonne numérique d'un tableau}
\usage{
enlever_valeurs_extremes(tableau, p_min, p_max)
}
\arguments{
\item{tableau}{Un dataframe contenant les données.}

\item{p_min}{Probabilité inférieure définissant la borne inférieure des valeurs à conserver.}

\item{p_max}{Probabilité supérieure définissant la borne supérieure des valeurs à conserver.}
}
\value{
Un dataframe avec les valeurs extrêmes remplacées par NA pour chaque colonne numérique.
}
\description{
Cette fonction élimine les valeurs qui sont en dehors des bornes définies par les probabilités \code{p_min} et \code{p_max}
pour chaque colonne numérique d'un tableau.
}
\examples{
# Exemple fictif pour illustrer l'utilisation :
df <- data.frame(a = c(1, 5, 1000, 10000, 20000), b = c(2, 10, 2000, 15000, 25000))
res <- enlever_valeurs_extremes(df, 0.00, 0.995)
print(res)

}
