% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_finess_fermes.R
\name{filtrer_finess_fermes}
\alias{filtrer_finess_fermes}
\title{Dans un tableau de données, avec un FINESS par ligne et les années en colonne, remplace les valeurs des années où le FINESS est fermé par NA}
\usage{
filtrer_finess_fermes(tab)
}
\arguments{
\item{tab}{tableau de données, une colonne "FINESS" et des colonnes d'années}
}
\value{
un tableau de données avec les valeurs des années où le FINESS est fermé remplacées par NA
}
\description{
Dans un tableau de données, avec un FINESS par ligne et les années en colonne, remplace les valeurs des années où le FINESS est fermé par NA
}
