% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_finess_fermes_camsp_cmpp.R
\name{filtrer_finess_fermes_camsp_cmpp}
\alias{filtrer_finess_fermes_camsp_cmpp}
\title{Dans un tableau de données, avec un FINESS par ligne et les années en colonne, remplace les valeurs des années où le FINESS n'est pas encore ouvert par NA}
\usage{
filtrer_finess_fermes_camsp_cmpp(tab, finess_full = charger_reference_finess())
}
\arguments{
\item{tab}{tableau de données, une colonne "FINESS" et des colonnes d'années}

\item{finess_full}{résultat de creer_reference_finess ou charger_reference_finess}
}
\value{
un tableau de données avec les valeurs des années où le FINESS n'est pas encore ouvert remplacées par NA
}
\description{
Dans un tableau de données, avec un FINESS par ligne et les années en colonne, remplace les valeurs des années où le FINESS n'est pas encore ouvert par NA
}
