% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_liste_a_vers_tablo.R
\name{format_liste_a_vers_tablo}
\alias{format_liste_a_vers_tablo}
\title{Format liste Années vers Liste variables}
\usage{
format_liste_a_vers_tablo(
  liste_a,
  var_liste = "ANNEE",
  format_sortie = "data.frame"
)
}
\arguments{
\item{liste_annees}{Une liste de dataframes, chaque élément de la liste représentant une année. La première colonne représente la variable identifiante}
}
\value{
Un tableau, avec une colonne "ANNE"
}
\description{
Cette fonction transforme une liste de variables (chaque élément de la liste = une variable, chaque colonne = une année)
en une liste de variables
}
