% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tablo_vers_liste_a.R
\name{format_tablo_vers_liste_a}
\alias{format_tablo_vers_liste_a}
\title{Convertir un tableau en une liste de tableaux élargis par année}
\usage{
format_tablo_vers_liste_a(
  tablo,
  variable_ident = "FINESS",
  variable_temporelle = "ANNEE",
  format_sortie = "data.frame"
)
}
\arguments{
\item{tablo}{Un tableau contenant au minimum les colonnes variable identifiante,une colonne variable temporelle et au moins une variable.}

\item{variable_temporelle}{variable avec le nom de la variable temporelle (exemple : ANNEE)}

\item{format_sortie}{data.frame par défaut, mais si autre chose sortira un data.table (format de travail de la fonction)}
}
\value{
Une liste de tableaux où chaque tableau représente une année
}
\description{
Cette fonction prend en entrée un tableau contenant des départements, des années et des variables,
et retourne une liste de tableaux où chaque élément représente une année
Dans chaque tableau, vous aurez une colonne variable identifiante et ensuite une colonne pour chaque variable
}
\examples{
\dontrun{
data <- tibble(
  DEPARTEMENT = c("01", "02", "03", "01", "02", "03", "01", "02", "03"),
  ANNEE = c(2014, 2014, 2014, 2015, 2015, 2015, 2016, 2016, 2016),
  VAR1 = c(10, 14, 16, 15, 19, 22, 10, 14, 25),
  VAR2 = c(NA, NA, NA, NA, NA, NA, 12, 18, 22)
)
res <- format_tablo_vers_liste_a(data, variable_ident = "DEPARTEMENT", variable_temporelle = "ANNEE", format_sortie = "data.frame")
}

}
