% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tablo_vers_liste_v.R
\name{format_tablo_vers_liste_v}
\alias{format_tablo_vers_liste_v}
\title{Convertir un tableau en une liste de tableaux élargis par variable}
\usage{
format_tablo_vers_liste_v(tablo, variable_ident = "DEPARTEMENT")
}
\arguments{
\item{tablo}{Un tableau contenant au minimum les colonnes variable identifiante, \code{ANNEE} et au moins une variable.}

\item{variable_ident}{variable avec le nom de la variable identifiante dans le tableau}
}
\value{
Une liste de tableaux où chaque tableau représente une variable.
}
\description{
Cette fonction prend en entrée un tableau contenant des départements, des années et des variables,
et retourne une liste de tableaux où chaque élément représente une variable.
Dans chaque tableau, vous aurez une colonne variable identifiante et ensuite une colonne pour chaque année.
}
\examples{
\dontrun{
data <- tibble(
  DEPARTEMENT = c("01", "02", "03", "01", "02", "03", "01", "02", "03"),
  ANNEE = c(2014, 2014, 2014, 2015, 2015, 2015, 2016, 2016, 2016),
  VAR1 = c(10, 14, 16, 15, 19, 22, 10, 14, 25),
  VAR2 = c(NA, NA, NA, NA, NA, NA, 12, 18, 22)
)
res <- format_tablo_vers_liste_v(data)
}

}
