% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france_depdom}
\alias{france_depdom}
\title{Carte de France métropolitaine + DOM}
\format{
A data frame with 101 rows and 3 columns
}
\source{
Package CARTElette
}
\usage{
france_depdom
}
\description{
Carte de France métropolitaine + DOM
}
\keyword{datasets}
