% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france_shapefile}
\alias{france_shapefile}
\title{Shapefile pour faire des cartes de France}
\format{
Un shapefile avec les colonnes suivantes :
\describe{
\item{code_insee}{Le code département}
\item{nom}{Le nom du département}
\item{nuts3}{Le code européen du département}
\item{wikipedia}{Le nom wikipedia}
\item{geometry}{les coordonnées du polygone}
}
}
\usage{
france_shapefile
}
\description{
Un shapefile qui contient les polygones par département de la France, y compris outre mer
}
\keyword{datasets}
