% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_creer_diagnostic_detaille.R
\name{imputation_creer_diagnostic_detaille}
\alias{imputation_creer_diagnostic_detaille}
\title{Imputation creer diagnostic détaillé}
\usage{
imputation_creer_diagnostic_detaille(
  base_stat,
  base_brute,
  var_reference,
  base_2019,
  categetab_sel = NULL
)
}
\arguments{
\item{base_stat}{une base stat au format tablo}

\item{base_brute}{une base brute au format tablo}

\item{base_2019}{la base 2019 de référence (contient PA/PH mais pas SAAD, avec imputation de capacité si pas présente)}

\item{categetab_sel}{argument optionnel, pour que le tableau ne sorte que sur certains categetab}
}
\value{
une liste d'années avec les résultats attendus pour chaque année
}
\description{
Principe : on a une base stat, une base brute, les deux au format tablo
On a une base 2019 pour les capacités
On regarde, année par année, et variable de la base par variable de la base, les taux de non réponse
}
