% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_format_annee_vers_variable.R
\name{imputation_format_annee_vers_variable}
\alias{imputation_format_annee_vers_variable}
\title{Change le format d'une table}
\usage{
imputation_format_annee_vers_variable(
  base_annee,
  noms_variable,
  nom_identifiant,
  noms_a_garder = NA
)
}
\arguments{
\item{base_annee}{Une liste de dataframes, un dataframe par année, contenant les mêmes variables.}

\item{noms_variable}{Un vecteur de caractères indiquant les noms des variables à traiter.}

\item{nom_identifiant}{Un caractère indiquant le nom de la variable identifiante.}

\item{noms_a_garder}{Un vecteur de caractères indiquant les noms des variables à conserver.
Par défaut à NA, signifiant qu'aucune autre variable n'est conservée.}
}
\value{
Une liste avec un tableau par variable contenant toutes les années.
}
\description{
Cette fonction transforme une liste de dataframes (un par année) avec les mêmes variables
en une liste avec un tableau par variable contenant toutes les années.
}
\examples{
# Exemple fictif pour illustrer l'utilisation :
# Supposons que base_annee est une liste contenant des dataframes pour 2020 et 2021.
base_2020 <- data.frame(id = 1:3, var1 = c(100, 200, 300), var2 = c("A", "B", "C"))
base_2021 <- data.frame(id = 1:3, var1 = c(110, 210, 320), var2 = c("A", "D", "E"))
base_annee <- list("2020" = base_2020, "2021" = base_2021)

# Utilisation de la fonction :
res <- imputation_format_annee_vers_variable(base_annee, c("var1", "var2"), "id")
print(res)

}
