% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_format_variable_vers_annee.R
\name{imputation_format_variable_vers_annee}
\alias{imputation_format_variable_vers_annee}
\title{Change le format d'une table de variables vers années}
\usage{
imputation_format_variable_vers_annee(
  base_variable,
  nom_identifiant,
  noms_a_garder = NA
)
}
\arguments{
\item{base_variable}{Une liste de dataframes, un dataframe par variable, contenant les mêmes années.}

\item{nom_identifiant}{Un caractère indiquant le nom de la variable identifiante.}

\item{noms_a_garder}{Un vecteur de caractères indiquant les noms des variables à conserver.
Par défaut à NA, signifiant qu'aucune autre variable n'est conservée.}
}
\value{
Une liste avec un tableau par année.
}
\description{
Cette fonction transforme une liste de dataframes (un par variable) avec les mêmes variables
en une liste avec un tableau par année.
}
\examples{
# Exemple fictif pour illustrer l'utilisation :
# Supposons que base_variable est une liste contenant des dataframes pour var1 et var2.
base_var1 <- data.frame(id = 1:3, "2020" = c(100, 200, 300), "2021" = c(110, 210, 320))
base_var2 <- data.frame(id = 1:3, "2020" = c("A", "B", "C"), "2021" = c("A", "D", "E"))
base_variable <- list("var1" = base_var1, "var2" = base_var2)

# Utilisation de la fonction :
res <- imputation_format_variable_vers_annee(base_variable, "id")
print(res)

}
