% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_interpolation.R
\name{imputation_interpolation}
\alias{imputation_interpolation}
\title{Imputation via Interpolation pour les Séries Temporelles}
\usage{
imputation_interpolation(df, nb_variables_de_cote = 0)
}
\arguments{
\item{df}{Un dataframe dont la première colonne est un identifiant. Les colonnes suivantes
doivent représenter des périodes temporelles consécutives.}

\item{nb_variables_de_cote}{Le nombre de colonnes (après l'identifiant) qui ne sont pas des
périodes temporelles et ne doivent pas être interpolées. Par défaut, cette valeur est 0.}
}
\value{
Un dataframe avec les valeurs manquantes interpolées linéairement.
}
\description{
Cette fonction effectue une interpolation linéaire pour les valeurs manquantes
dans un dataframe contenant des séries temporelles.
}
\examples{
\dontrun{
df <- data.frame(id = 1:5,
                 t1 = c(NA, 1, 2, NA, 5),
                 t2 = c(NA, NA, 3, 4, NA),
                 t3 = c(NA, 3, NA, 5, 7))
imputed_df <- imputation_interpolation(df)
}

}
