% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_modalite_frequente.R
\name{imputation_modalite_frequente}
\alias{imputation_modalite_frequente}
\title{Imputation de la Modalité la Plus Fréquente}
\usage{
imputation_modalite_frequente(
  tablo,
  variable_a_imputer,
  variable_de_groupe = NA
)
}
\arguments{
\item{tablo}{Un dataframe contenant la variable à imputer ainsi que la variable de groupe (si spécifiée).}

\item{variable_a_imputer}{Le nom (sous forme de chaîne de caractères) de la variable du dataframe dont les valeurs \code{NA} doivent être imputées.}

\item{variable_de_groupe}{Le nom (sous forme de chaîne de caractères) de la variable de groupe à utiliser pour l'imputation segmentée. Par défaut, \code{NA}, ce qui signifie qu'aucune variable de groupe n'est utilisée.}
}
\value{
Un dataframe avec les valeurs \code{NA} imputées dans la variable spécifiée.
}
\description{
Cette fonction remplace les valeurs manquantes (\code{NA}) dans une colonne spécifiée d'un tableau par la modalité la plus fréquente.
L'utilisateur peut également spécifier une variable de groupe pour effectuer cette imputation de manière segmentée.
}
\examples{
df <- data.frame(
  groupe = c("A", "A", "B", "B", "C", "C"),
  valeur = c(1, NA, 2, 2, 3, NA)
)
imputation_modalite_frequente(df, "valeur", "groupe")

}
