% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_taux_croissance.R
\name{imputation_taux_croissance}
\alias{imputation_taux_croissance}
\title{Imputation de Valeurs Manquantes par Taux de Croissance}
\usage{
imputation_taux_croissance(
  tablo,
  table_reference = NA,
  variable_groupe = NA,
  affichage = FALSE,
  taux_croissance = NULL
)
}
\arguments{
\item{tablo}{Un dataframe avec une colonne identifiante suivie de colonnes temporelles.}

\item{table_reference}{Un dataframe optionnel contenant des données de référence pour le regroupement.}

\item{variable_groupe}{Un caractère indiquant la colonne de regroupement utilisée dans la table de référence.}

\item{affichage}{Un booléen pour activer ou désactiver l'affichage des informations sur les valeurs manquantes.}
}
\value{
Un dataframe avec les valeurs manquantes imputées.
}
\description{
Impute les valeurs manquantes d'une série temporelle en utilisant un taux de croissance.
Les séries temporelles sont groupées par une variable de groupe, et le taux de croissance
est appliqué de manière chronologique puis antéchronologique.
}
\examples{
\dontrun{
df <- data.frame(id = c(1,2,3),
                 annee1 = c(100, NA, 130),
                 annee2 = c(NA, 120, 140))
df_reference <- data.frame(id = c(1,2,3), groupe = c('A','A','B'))
df_imputed <- imputation_taux_croissance(df, table_reference = df_reference,
variable_groupe = "groupe", affichage = TRUE)
}

}
