% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_totale.R
\name{imputation_totale}
\alias{imputation_totale}
\title{Imputation de Valeurs Manquantes par Ratio de Référence}
\usage{
imputation_totale(
  tablo,
  nom_a_imputer,
  nom_identifiant,
  table_de_reference,
  variables_de_groupe,
  variable_de_ratio
)
}
\arguments{
\item{tablo}{Un dataframe contenant la variable à imputer.}

\item{nom_a_imputer}{Nom de la colonne de \code{tablo} contenant les valeurs à imputer.}

\item{nom_identifiant}{Nom de la colonne servant de clé pour la jointure avec \code{table_de_reference}.}

\item{table_de_reference}{Un dataframe contenant les données de référence pour le calcul du ratio.}

\item{variables_de_groupe}{Noms des colonnes pour regrouper les données lors du calcul du ratio.}

\item{variable_de_ratio}{Nom de la colonne utilisée comme dénominateur lors du calcul du ratio.}
}
\value{
Un vecteur avec les valeurs imputées.
}
\description{
Impute les valeurs manquantes d'une variable en utilisant un ratio calculé à partir
de données de référence. Le ratio est calculé en regroupant les données par certaines
variables et en divisant la somme de la variable cible par la somme d'une autre variable.
}
\examples{
\dontrun{
df <- data.frame(id = c(1,2,3),
                 valeur = c(100, NA, 130))
df_reference <- data.frame(id = c(1,2,3),
                           groupe = c('A','A','B'),
                           ratio_var = c(10, 20, 30))
imputed_values <- imputation_totale(df, "valeur", "id", df_reference, "groupe", "ratio_var")
}

}
