% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne_filtree.R
\name{moyenne_filtree}
\alias{moyenne_filtree}
\title{Moyenne filtrée}
\usage{
moyenne_filtree(vecteur)
}
\arguments{
\item{vecteur}{un vecteur de données numériques}
}
\value{
un numérique avec la moyenne filtrée
}
\description{
Cette fonction propose de calculer la moyenne filtrée d'un vecteur, i.e.
la moyenne sans les valeurs min et max du vecteur
}
