% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_left.R
\name{pad_left}
\alias{pad_left}
\title{pad_left}
\usage{
pad_left(x, len = 1 + max(nchar(x)), char = "0")
}
\arguments{
\item{x}{chaine de caractère}

\item{len}{longueur cible de la sortie}

\item{char}{caractère à insérer pour compléter}
}
\value{
renvoie une chaine de caractères
}
\description{
pad_left
}
\examples{
pad_left("bonjour",10,"0")
}
