% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_geometry.R
\name{place_geometry}
\alias{place_geometry}
\title{Place a Geometry at a Specified Position}
\usage{
place_geometry(geometry, position, scale = 1)
}
\arguments{
\item{geometry}{A \code{sf} geometry object that you wish to move.}

\item{position}{A numeric vector of length 2 specifying the x and y coordinates to move the geometry's centroid to.}

\item{scale}{A numeric value (default is 1) by which to scale the geometry.}
}
\value{
A \code{sf} geometry object that has been moved and scaled as specified.
}
\description{
This function moves an existing geometry (\code{geometry}) to a given position (\code{position}) and scales it by a specified factor (\code{scale}).
}
