% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrouper_categories.R
\encoding{UTF-8}
\name{regrouper_categories}
\alias{regrouper_categories}
\title{Cette fonction permet de transformer une catégorie d'ESMS détaillé en une catégorie}
\usage{
regrouper_categories(tableau, nom_var_cat)
}
\arguments{
\item{tableau}{un tableau sous format dataframe incluant le statut juridique parmi les colonnes}

\item{nom_var_cat}{le nom de la colonne incluant la catégorie (categetab a priori)}
}
\value{
le tableau initial, complété par une colonne "categetab_regr" incluant le statut juridique regroupé
}
\description{
Cette fonction permet de transformer une catégorie d'ESMS détaillé en une catégorie
}
\examples{
df_exemple <- data.frame(categetab = c(183, 186, 188, 255), autre_colonne = c(10, 23, 47, 54))
resultat <- regrouper_categories(df_exemple, "categetab")
print(resultat)
}
