% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrouper_statut_juridique.R
\encoding{UTF-8}
\name{regrouper_statut_juridique}
\alias{regrouper_statut_juridique}
\title{Cette fonction permet de transformer un statut juridique détaillé en un statut juridique regroupé}
\usage{
regrouper_statut_juridique(tableau, nom_var_stat)
}
\arguments{
\item{tableau}{un tableau sous format dataframe incluant le statut juridique parmi les colonnes}

\item{nom_var_stat}{le nom de la colonne incluant le statut juridique}
}
\value{
le tableau initial, complété par une colonne "statut_jur_fin" incluant le statut juridique regroupé
}
\description{
Cette fonction permet de transformer un statut juridique détaillé en un statut juridique regroupé
}
\examples{
df_exemple <- data.frame(stat_juridique = c("01", "10", "40", "70"),
autre_colonne = c(10, 23, 47, 54))
resultat <- regrouper_statut_juridique(df_exemple, "stat_juridique")
print(resultat)
}
