% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_finess_dominant.R
\name{selection_finess_dominant}
\alias{selection_finess_dominant}
\title{Selection finess dominant}
\usage{
selection_finess_dominant(base_finess_reduite, variable, seuil)
}
\arguments{
\item{base_finess_reduite}{base finess reduite stockée sur GEOD pour une année}

\item{variable}{le nom de la variable à tester}

\item{seuil}{le seuil choisi pour le test}
}
\value{
un tableau précisant par FINESS si ce dernier a une modalité dominante
}
\description{
Cette fonction part d'une base finess_reduite (avec toutes les années), et cela évalue si
le finess sélectionné a une part de places supérieure au seuil
}
\details{
Exemple : seuil = 80\%, variable = "client" : on vérifie si le FINESS a plus de 80\% de ses places sur un type de déficience (client)
}
