% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcer_fichiers.R
\name{sourcer_fichiers}
\alias{sourcer_fichiers}
\title{Source l'ensemble des fichiers R d'un répertoire donné}
\usage{
sourcer_fichiers(repertoire)
}
\arguments{
\item{repertoire}{adresse du répertoire contenant les fichiers}
}
\value{
rien du tout, si ce n'est des messages dans la console
}
\description{
Source l'ensemble des fichiers R d'un répertoire donné
}
