% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirer_selon_loi_normale.R
\name{tirer_selon_loi_normale}
\alias{tirer_selon_loi_normale}
\title{Fonction qui tire selon une loi normale}
\usage{
tirer_selon_loi_normale(vecteur, nb_a_tirer, tests = FALSE, positif = TRUE)
}
\arguments{
\item{vecteur}{vecteur de variables (numériques) de référence}

\item{nb_a_tirer}{nb d'éléments à tirer selon la loi normale}

\item{tests}{TRUE ou FALSE, selon qu'on veut afficher les résultats des tests de normalité (facultatif ; par défaut : FALSE)}

\item{positif}{TRUE ou FALSE selon qu'on accepte ou non les valeurs négatives. Par défaut, TRUE (le résultat ne doit inclure que des valeurs positives)}
}
\value{
un vecteur de taille nb_a_tirer tiré suivant une loi normale qui suit la même que l'élément "vecteur" en entrée
}
\description{
Cette fonction permet de tirer aléatoirement des variables qui suivent une loi normale identique à celle calculée pour une variable de référence.
Pour tester si l'hypothèse de normalité est plausible, il est possible d'utiliser l'argument tests = TRUE
}
