% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_variable.R
\name{transformer_variable}
\alias{transformer_variable}
\title{Cette fonction transforme une variable numérique en une variable catégorielle}
\usage{
transformer_variable(
  variable_entree,
  type_souhaite = "CAT_AUTO",
  classes_souhaitees = 5
)
}
\arguments{
\item{variable_entree}{vecteur au format idéalement numérique}

\item{type_souhaite}{type de sortie souhaité de la variable. Par défaut "CAT_AUTO", c'est à dire variable catégorielle avec des seuils calculés automatiquement}

\item{classes_souhaitees}{Nombre de classes souhaitées (si on est dans de l'auto) ou vecteur de classes (si on est dans du manuel)}
}
\value{
une variable modifiée pour devenir catégorielle (a priori)
}
\description{
Ceci ne fonctionne que si la variable d'entrée est numérique, sinon on se contente de renvoyer la même
}
