% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_carte_indiv.R
\name{creer_carte_indiv}
\alias{creer_carte_indiv}
\title{Creer carte individuelle}
\usage{
creer_carte_indiv(
  donnees,
  region = "FRANCEMETRO",
  palette,
  titre_legende = "Legende",
  afficher_valeurs = FALSE,
  couleur_valeurs = "black",
  arrondi_valeurs = NA,
  taille_valeurs = 2,
  afficher_legende = TRUE,
  formatter_valeurs = NULL,
  save_png = FALSE
)
}
\arguments{
\item{donnees}{un tableau de données, format SF, avec les coordonnées}

\item{region}{la région souhaitée, incluse dans le SF (a priori FRANCEMETRO, 971, 972,...)}

\item{palette}{une palette de couleurs, fournie a priori avant (note : peut être à changer)}

\item{titre_legende}{Si la légende s'affiche : qu'est-il marqué au dessus ?}

\item{afficher_valeurs}{TRUE ou FALSE : détermine si on souhaite afficher les valeurs choisies sur la carte. Par défaut : FALSE}

\item{couleur_valeurs}{couleur d'affichage des valeurs sur la carte. par défaut "black".}

\item{arrondi_valeurs}{Si on souhaite arrondir les valeurs, par exemple au milliers. Le principe est celui de round, dans R : round(115,4554, 1) => 115,5 / round(115,45554, -2) => 100}

\item{taille_valeurs}{taille des valeurs affichées, en pourcentage de la hauteur de la carte. Par défaut, 2 (pour 2\%)}

\item{afficher_legende}{affiche, ou non, la légende à côté de la carte}

\item{formatter_valeurs}{Fonction de formattage des valeurs du tooltip.}

\item{save_png}{TRUE ou FALSE, selon qu'on souhaite ou non sauver un png avec l'image}
}
\value{
une carte format leaflet
}
\description{
Cette fonction crée une carte pour une région donnée (FRANCEMETRO, ou un département d'outre mer)
}
