% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_reference_finess.R
\name{creer_reference_finess}
\alias{creer_reference_finess}
\title{Créer référence finess}
\usage{
creer_reference_finess(origine = "GEOD", annee_ref = 2019)
}
\value{
une liste d'années
une base avec la dernière version de chaque finess
un élément synthèse avec une valeur de référence pour chaque
un tableau des capacités pour chaque année
un tableau 0-1 des finess fermés-ouverts (capinsTOT à NA ou pas) chaque année
un tableau de la démographique des structures par catégorie
}
\description{
Crée une base de référence finess, utilisable pour du redressement, ou pour toute autre chose et l'enregistre
}
