% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_toutes_cartes.R
\name{creer_toutes_cartes}
\alias{creer_toutes_cartes}
\title{Cette fonction lance la création des différentes cartes utiles (notamment DOM)}
\usage{
creer_toutes_cartes(france_sf, palette, titre_legende)
}
\arguments{
\item{france_sf}{Un tableau de données SF, qui inclut notamment les coordonnées des départements (polygones)}

\item{palette}{une palette de couleurs}

\item{titre_legende}{Le nom de la légende qui doit apparaitre}
}
\value{
Beaucoup de cartes, stockées au format png (une par département)
}
\description{
Dans le principe, cette fonction lance la création d'une carte par région de la base, et la sauve en png
Cette sauvegarde se fait dans le répertoire courant (option mettable pour plus de flexibilité)
}
