% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faire_une_carte.R
\name{faire_une_carte}
\alias{faire_une_carte}
\title{Fonction principale pour faire une carte}
\usage{
faire_une_carte(
  table,
  type_output = "image",
  type_visu = "FRANCE_ENTIERE",
  titre_legende = "Legende",
  type_var_souhait = "CAT_AUTO",
  classes_souh = 5,
  chemin_sortie = "",
  type_palette = "viridis"
)
}
\arguments{
\item{table}{Un tableau de données. Deux colonnes : la première avec des départements, la 2e avec la variable à mettre sur la carte (c'est la seule contrainte pour que cela marche !)}

\item{type_output}{Type d'output souhaité (à terme, doit permettre d'intégrer une option "shiny")}

\item{type_visu}{Type de visualisation. Par défaut "France Entière", c'est à dire y compris outremer}

\item{titre_legende}{Le nom de la légende (par défaut "Legende")}

\item{type_var_souhait}{Le type de variable souhaité. Cela peut être "CAT_AUTO" : la variable est alors automatiquement transformée en catégorielle si besoin. "NUM" : dans ce cas, si la variable est numérique, elle reste telle quelle.}

\item{classes_souh}{Classes souhaitées. Si on est dans la CAT_AUTO, il faut un nombre qui est le nombre de classes ; si on est dans CAT_MAN : alors on fournit les catégories. Par défaut CAT_AUTO}

\item{chemin_sortie}{le répertoire dans lequel on enregistre le jpg résultat (par défaut le répertoire de travail)}

\item{type_palette}{Type de la palette (par défaut : viridis)}
}
\value{
une carte format jpg
}
\description{
Fournit pas mal d'options (en construction :)) pour obtenir des cartes dans différentes circonstances
Le principe est d'utiliser un leaflet, voire plusieurs si on veut afficher les DOM
}
