% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_liste_a_vers_liste_v.R
\name{format_liste_a_vers_liste_v}
\alias{format_liste_a_vers_liste_v}
\title{Format liste Années vers Liste variables}
\usage{
format_liste_a_vers_liste_v(liste_annees)
}
\arguments{
\item{liste_annees}{Une liste de dataframes, chaque élément de la liste représentant une année. La première colonne représente la variable identifiante}
}
\value{
Une liste de dataframes, cette fois chaque élément représente une variable
}
\description{
Cette fonction transforme une liste d'années (chaque élément de la liste = une année, chaque colonne = une variable)
en une liste de variables
}
