% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_liste_v_vers_tablo.R
\name{format_liste_v_vers_tablo}
\alias{format_liste_v_vers_tablo}
\title{Transforme une liste de variables en un tableau}
\usage{
format_liste_v_vers_tablo(liste_v)
}
\arguments{
\item{liste_v}{une liste de variables}
}
\value{
un tablo, i.e. une grand tableau avec en première colonne la variable identifiante, en 2e les années (sur plusieurs lignes), puis les autres colonnes sont les variables
}
\description{
Liste de variables = liste de tableaux, première colonne = variable identifiante, les suivantes = des années
}
